ocr-service/requirements.txt
# Web Server
fastapi>=0.136.0
uvicorn>=0.30.0

# OCR and Document Processing
docling>=2.90.0

# Environment and Utilities
python-dotenv>=1.0.1
python-multipart>=0.0.12

# Mapping Agent (Pydantic)
pydantic>=2.12.0
pydantic-ai[google]>=1.67.0

# FHIR
fhir.resources>=7.1.0
